Simplified argument handling code.
An empty Vec will not allocate until an element is pushed on to it, and
there is nu functional difference between None and Some(vec![]), so
`Option<Vec<u8>>` was replaced with just `Vec<u8>`.

Also changed `encode_message` to build a type tag String directly,
instead of collecting chars, to avoid an allocation/encoding to convert
it back to a String. This also allows `encode_string_into` to be used
there, meaningn `encode_string` is no longer used internally (but is
kept since it is in the public API).