Implemented an `Output` trait in `encoder`.
This trait can be applied to anything which should be able to receive
OSC output. This allows for flexible output, including but not limited
to advanced buffer usage, streaming output, etc.

It comes with an implementation for `Vec<u8>`, in which all writes
append to the Vec. Implementations for `Cursor` and
`Seek + Write` from `std` to follow.